Test if monitor_int is defined

+++ lib/pf/services.pm
@@ -405,13 +405,15 @@ sub service_ctl {
                 }
                 elsif ($daemon =~ "snort") {
                     $pid = 0;
-                    if (-e "$install_dir/var/run/${daemon}_${monitor_int}.pid") {
-                        chomp( $pid = `cat $install_dir/var/run/${daemon}_${monitor_int}.pid`);
-                        my $ppt = new Proc::ProcessTable;
-                        my $proc = first { defined($_) } grep { $_->pid == $pid } @{ $ppt->table };
-                        if (!defined($proc)) {
-                            unlink( $install_dir . "/var/run/${daemon}_${monitor_int}.pid" );
-                            return(0);
+                    if (defined $monitor_int) {
+                        if (-e "$install_dir/var/run/${daemon}_${monitor_int}.pid") {
+                            chomp( $pid = `cat $install_dir/var/run/${daemon}_${monitor_int}.pid`);
+                            my $ppt = new Proc::ProcessTable;
+                            my $proc = first { defined($_) } grep { $_->pid == $pid } @{ $ppt->table };
+                            if (!defined($proc)) {
+                                unlink( $install_dir . "/var/run/${daemon}_${monitor_int}.pid" );
+                                return(0);
+                            }
                         }
                     }
                     return ($pid);
